+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 ## Fixed
 - allow scanning of certain qr code types on welcome screen (account, url and text)
 - fix selecting chat after forwarding to it
+- fixed bug where unread badge on app icon was not updated immediately
+- fix notifications are not removed from notification center
 - fix quote linebreaks #2870
 - fix low resolution of copy qrcode image #2907